update functionality added

+++ src/Pages/MyReviews/MySingleReview.js
@@ -1,20 +1,27 @@
 import React from 'react';
+import { Link } from 'react-router-dom';
+
 
 const MySingleReview = ({ review, setRefresh, refresh }) => {
-    console.log(review);
+
 
     const handleDelete = (id) => {
-        fetch(`http://localhost:5000/delete/${id}`, {
-            method: 'DELETE'
-        })
-            .then(res => res.json())
-            .then(result => {
-                if (result) {
-                    setRefresh(!refresh);
-                    alert('Review Deleted Successfully')
-                }
+        const proceed = window.confirm('Are you sure, you want to delete this review?');
+
+        if (proceed) {
+            fetch(`http://localhost:5000/delete/${id}`, {
+                method: 'DELETE'
             })
+                .then(res => res.json())
+                .then(result => {
+                    if (result) {
+                        setRefresh(!refresh);
+                        alert('Review Deleted Successfully')
+                    }
+                })
+        }
     }
+
     const { date, details, ratings, title, _id } = review;
     return (
         <tbody>
@@ -33,7 +40,7 @@ const MySingleReview = ({ review, setRefresh, refresh }) => {
                 <td>{ratings}</td>
                 <th>
                     <button onClick={() => handleDelete(_id)} className="btn btn-ghost btn-xs">Delete</button>
-                    <button className="btn btn-ghost btn-xs">Edit</button>
+                    <button className="btn btn-ghost btn-xs"><Link to={`/update-review/${_id}`}>Edit</Link></button>
                 </th>
             </tr>
         </tbody>
